[ios] Add back support for annotation interactions
Since the previous implementation always assumed it was dealing with an
image annotation, the introduction of UIViews as annotations
broke interactions with annotations. For example, callouts
would no longer appear over annotations when they otherwise should
since the map no longer believed the annotation (as a UIView) was
visible on the map.

This updates the hit test logic in the appropriate places on the map
so that interactions with annotations (i.e. selecting an annotation
work again). This also works for recent work done to move an
annotation in the view when the annotation model location changes.

Finally, a less trivial MGLAnnotationView was added for testing.