[ios] Introduce UIView annotation view reuse
Add a reuse queue implementation in MGLMapView. Delegates can now
call `dequeueReusableAnnotationViewWithIdentifier:` to attempt to
obtain a previously used annotation view. Annotation views are
stored on the queue and to be reuse when they go offscreen.

The context object's reference to an annotation view is updated
to always point to the active view onscreen, even if the view
has gone on and off screen again. The delegate must provide a new
view if no view is available to reuse.