[ios] Make annotation view updates and reuse more efficient
This removes a loop (and space complexity) of all annotation views
to hide them before working out annotation updates and reuse
work. Instead of looping through all views, the visible tags are stored
in a set and looked up in the second loop and used in that loop to
help determine reuse eligibility.

In addition, the annotation view update and reuse logic has been
moved to a new internal method that is now triggered in
`notifyMapChange` for the case of `MapChangeRegionIsChanging`.

The iosapp example annotation view is refactored to use a nib
instead of a drawRect override - this would be a best practice
for any client that wants better performance when a non trivial
amount of UIView annotations are displayed.