[ios] Guard against annotation reuse when there is no identifier
This protects against a crash that would occur when view reuse was
attempted in cases where the map view delegate did not ask for it.
It also fixes a bug where annotation views were attempted when
the map view delegate did not implement the correct delegate method.

It also moves the call to `updateAnnotationViews` back to
`glkView:drawInRect` since doing it in `notifyMapChange`
introduced lag. Additional optimizations may occur in future commits
to make the work required to hide and show views and handle view
reuse less intensive.

In addition, the example MBXAnnotationView is simplified to just
a programatic view with a center panel that can change color. The
color changing center panel is used to indicate if a vie has
gone through the reuse queue.