[ios] Make annotation views selectable
This adds a tap gesture recognizer to the MGLAnnotationView base class.
Any tap of the annotation view triggers a notification to the
annotation view delegate. In practice the only delegate of the
annotation view is the map view.

The map view has been modified to
respond to annotation view tap gesture events by selecting the
annotation represented by the view. This triggers a callout if the
map view delegate allows callouts to be shown.

This also fixes a bug where, if `selectAnnotation:` was called for
an annotation that was not previously added, the annotationTag was
not captured after `selectAnnotation:` called `addAnnotation:`. This
caused the annotation to not be selected even after it was added.