[ios] Add access token plist and annotations example to playground
This updates the playground to look for a plist with an access
token. The plist is ignored by git so there is less chance of
mistakenly committing an access token in the playground file.

In addition, an example of adding view based annotations and
interacting with those annotations is added.